third_party/blink/web_tests/virtual/threaded/animations/KeyframeEffect-composited-animation.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animating with KeyframeEffect objects</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<div id="target">x</div>
<script>
"use strict";

promise_test(function(t) {
  var effect = new KeyframeEffect(target, { opacity: [0, 0.9] }, 1000);
  var anim = target.animate(null);
  anim.effect = effect;

  return Promise.all([anim.ready]).then(function() {
    assert_true(internals.isCompositedAnimation(anim),
        "Opacity animation with KeyframeEffect should be composited");
  });
}, "Using KeyframeEffect should not change whether an animation is composited");

</script>
</body>
